Completed creating, deleting news and updating ckeditor version

+++ TDTU-IT-alumni-management-system/Areas/Admin/Controllers/NewsController.cs
@@ -6,6 +6,7 @@ using System.Data.Entity.Validation;
 using System.IO;
 using System.Linq;
 using System.Net;
+using System.Text;
 using System.Web;
 using System.Web.Mvc;
 using TDTU_IT_alumni_management_system.Models;
@@ -23,7 +24,7 @@ namespace TDTU_IT_alumni_management_system.Areas.Admin.Controllers
         }
 
         // GET: Admin/News/Details/5
-        public ActionResult Details(string id)
+        public ActionResult Details(int? id)
         {
             if (id == null)
             {
@@ -40,6 +41,17 @@ namespace TDTU_IT_alumni_management_system.Areas.Admin.Controllers
         // GET: Admin/News/Create
         public ActionResult Create()
         {
+            // Kiểm tra xem có thông báo lỗi trong TempData không
+            if (TempData["ErrorMessage"] != null)
+            {
+                // Lấy dữ liệu đã nhập từ TempData để hiển thị lại trên giao diện
+                var formData = TempData["FormData"] as News;
+
+                // Truyền dữ liệu đã lưu vào mô hình để hiển thị lại trên giao diện
+                return View(formData);
+            }
+
+            // Nếu không có thông báo lỗi, trả về trang Create bình thường
             return View();
         }
 
@@ -49,7 +61,7 @@ namespace TDTU_IT_alumni_management_system.Areas.Admin.Controllers
         [HttpPost]
         [ValidateAntiForgeryToken]
         [ValidateInput(false)]
-        public ActionResult Create([Bind(Include = "IDNews,Title,Content,ImgNews,meta,hide,order,datebegin")] News news , HttpPostedFileBase ImgNews)
+        public ActionResult Create([Bind(Include = "IDNews,Title,Description,Content,ImgNews,meta,hide,order,datebegin")] News news , HttpPostedFileBase img)
         {
             try
             {
@@ -57,39 +69,54 @@ namespace TDTU_IT_alumni_management_system.Areas.Admin.Controllers
                 var filename = "";
                 if (ModelState.IsValid)
                 {
-                    if (ImgNews != null)
+                    // Kiểm tra xem các trường bắt buộc có được nhập hay không
+                    if (string.IsNullOrEmpty(news.Title) || string.IsNullOrEmpty(news.Description) || string.IsNullOrEmpty(news.Content) || img == null)
                     {
-                        filename = DateTime.Now.ToString("dd-MM-yy-hh-mm-ss-") + ImgNews.FileName;
-                        path = Path.Combine(Server.MapPath("~/Upload/images/News/"), filename);
-                        ImgNews.SaveAs(path);
+                        string errorMessage = "Vui lòng nhập đầy đủ thông tin cho các trường bắt buộc.";
+                        TempData["ErrorMessage"] = errorMessage;
+
+                        // Lưu trữ dữ liệu đã nhập vào TempData để khôi phục sau khi đóng modal
+                        TempData["FormData"] = news;
+
+                        return RedirectToAction("Create");
+                    }
+                    if (img != null)
+                    {
+                        filename = DateTime.Now.ToString("dd-MM-yy-hh-mm-ss-") + img.FileName;
+                        path = Path.Combine(Server.MapPath("~/Upload/images/News"), filename);
+                        img.SaveAs(path);
                         news.ImgNews = filename; //Lưu ý
                     }
                     else
                     {
                         news.ImgNews = "logo.png";
                     }
-                    news.IDNews = "10";
                     news.datebegin = Convert.ToDateTime(DateTime.Now.ToShortDateString());
-                    //news.meta = ConvertToUnSign(news.Title); //convert Tiếng Việt không dấu
                     db.News.Add(news);
                     db.SaveChanges();
                     return RedirectToAction("Index");
                 }
             }
-            catch (DbEntityValidationException e)
+            catch (DbEntityValidationException ex)
             {
-                throw e;
+                foreach (var validationErrors in ex.EntityValidationErrors)
+                {
+                    foreach (var validationError in validationErrors.ValidationErrors)
+                    {
+                        Console.WriteLine("Property: {0} Error: {1}",
+                            validationError.PropertyName, validationError.ErrorMessage);
+                    }
+                }
+                // Nếu có lỗi, thiết lập thông báo lỗi và lưu trữ dữ liệu đã nhập vào TempData
+                TempData["ErrorMessage"] = "Có lỗi xảy ra khi tạo tin tức. Vui lòng kiểm tra lại thông tin.";
+                TempData["FormData"] = news;
+                return RedirectToAction("Create");
             }
-            catch (Exception ex)
-            {
-                throw ex;
-            }
-
             return View(news);
         }
 
         // GET: Admin/News/Edit/5
-        public ActionResult Edit(string id)
+        public ActionResult Edit(int? id)
         {
             if (id == null)
             {
@@ -108,38 +135,72 @@ namespace TDTU_IT_alumni_management_system.Areas.Admin.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult Edit([Bind(Include = "IDNews,Title,Content,ImgNews,meta,hide,order,datebegin")] News news)
+        [ValidateInput(false)]
+        public ActionResult Edit([Bind(Include = "IDNews,Title,Content,ImgNews,meta,hide,order,datebegin")] News news , HttpPostedFileBase img)
         {
-            if (ModelState.IsValid)
+            try
             {
-                db.Entry(news).State = EntityState.Modified;
-                db.SaveChanges();
-                return RedirectToAction("Index");
+                var path = "";
+                var filename = "";
+                News temp = getById(news.IDNews);
+                if (ModelState.IsValid)
+                {
+                    if (img != null)
+                    {
+                        filename = DateTime.Now.ToString("dd-MM-yy-hh-mm-ss-") + img.FileName;
+                        path = Path.Combine(Server.MapPath("~/Upload/images/News"), filename);
+                        img.SaveAs(path);
+                        temp.ImgNews = filename; //Lưu ý
+                    }
+                    temp.Title = news.Title;
+                    temp.Content = news.Content;
+                    temp.Description = news.Description;
+                    news.datebegin = Convert.ToDateTime(DateTime.Now.ToShortDateString());
+                    temp.meta = ConvertToUnSign(news.meta); //convert Tiếng Việt không dấu
+                    db.Entry(temp).State = EntityState.Modified;
+                    db.SaveChanges();
+                    return RedirectToAction("Index");
+                }
+            }
+            catch (DbEntityValidationException e)
+            {
+                throw e;
+            }
+            catch (Exception ex)
+            {
+                throw ex;
             }
             return View(news);
         }
 
         // GET: Admin/News/Delete/5
-        public ActionResult Delete(string id)
+        public ActionResult Delete(int? id)
         {
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
+
             News news = db.News.Find(id);
             if (news == null)
             {
                 return HttpNotFound();
             }
-            return View(news);
+
+            return PartialView("_DeleteConfirmation", news);
         }
 
         // POST: Admin/News/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
-        public ActionResult DeleteConfirmed(string id)
+        public ActionResult DeleteConfirmed(int? id)
         {
             News news = db.News.Find(id);
+            if (news == null)
+            {
+                return HttpNotFound();
+            }
+
             db.News.Remove(news);
             db.SaveChanges();
             return RedirectToAction("Index");
@@ -153,20 +214,40 @@ namespace TDTU_IT_alumni_management_system.Areas.Admin.Controllers
             }
             base.Dispose(disposing);
         }
+        // Hàm chuyển đổi dấu thành không dấu và loại bỏ các ký tự không mong muốn
         private string ConvertToUnSign(string input)
         {
-            // Bảng chữ cái có dấu và không dấu tương ứng
-            string[] signs = new string[] { "aAeEoOuUiIdDyY", "áàạảãâấầậẩẫăắằặẳẵÁÀẠẢÃÂẤẦẬẨẪĂẮẰẶẲẴéèẹẻẽêếềệểễÉÈẸẺẼÊẾỀỆỂỄóòọỏõôốồộổỗơớờợởỡÓÒỌỎÕÔỐỒỘỔỖƠỚỜỢỞỠúùụủũưứừựửữÚÙỤỦŨƯỨỪỰỬỮíìịỉĩÍÌỊỈĨđĐýỳỵỷỹÝỲỴỶỸ" };
+            StringBuilder result = new StringBuilder();
 
-            // Duyệt qua mỗi ký tự trong input
-            for (int i = 0; i < signs[0].Length; i++)
+            foreach (char c in input)
             {
-                // Thay thế ký tự có dấu bằng ký tự không dấu
-                input = input.Replace(signs[1][i], signs[0][i]);
+                // Kiểm tra nếu ký tự là chữ cái, chữ số hoặc dấu gạch ngang
+                if (char.IsLetterOrDigit(c) || c == '-')
+                {
+                    // Thêm ký tự vào chuỗi kết quả
+                    result.Append(c);
+                }
             }
 
-            // Trả về kết quả
-            return input;
+            return result.ToString();
         }
+        public News getById(long id)
+        {
+            return db.News.Where(x => x.IDNews == id).FirstOrDefault();
+
+        }
+        //private string GetFirstWords(string input, int wordCount)
+        //{
+        //    // Tách chuỗi thành mảng các từ
+        //    string[] words = input.Split(new char[] { ' ' }, StringSplitOptions.RemoveEmptyEntries);
+
+        //    // Lấy số từ cần thiết từ mảng
+        //    int numWords = Math.Min(wordCount, words.Length);
+
+        //    // Kết hợp các từ thành chuỗi kết quả
+        //    string result = string.Join(" ", words, 0, numWords);
+
+        //    return result;
+        //}
     }
 }
